requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.1.5
django-jazzmin==3.0.1
djangorestframework==3.15.2
iniconfig==2.0.0
isort==5.13.2
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.4
sqlparse==0.5.3
typing_extensions==4.12.2
